service_registry/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-dirtyfields==1.9.2
djangorestframework==3.15.1
drf-nested-routers==0.93.5
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2024.1